GR0021 - Correção de Layout

+++ Back-End/bovinainfo/src/main/java/com/gestaorural/bovinainfo/model/Funcionario.java
@@ -8,8 +8,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
@@ -19,7 +17,7 @@ public class Funcionario {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;	
+	private Long id;	
 
 	@Column(name = "nome")
 	private String nome;
@@ -37,11 +35,11 @@ public class Funcionario {
 	private Endereco endereco;
 	
 	// Getters and Setters
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
